correct validation of simple content restriction of complex types
The documentation for the "restriction" element in the SfXS states that
base attribute and simpleType child are mutually exclusive.  This is
backed up by http://www.w3.org/TR/xmlschema-2/#defn-rep-constr.

However: This constraint only applies when the "restriction" element is
used in simpleType definitions; it is not true for complexType
definition with simple content, which specifically allow simpleType to
appear in a restriction when the base has mixed emptiable content.

See http://lists.xml.org/archives/xml-dev/200103/msg00601.html and
hasty follow-up; unfortunately the comment was not fixed.

Correct the RelaxNG schema to distinguish the restriction element within
a simpleType definition from the one within a simpleContent element.