correct annotation sub-element support for wildcard attributes
documentation and appinfo elements permit unrecognized attributes as
long as they aren't in the XSD namespace.  Unfortunately adding this
requires that we hide the fact that a source attribute in no namespace
is explicitly supported.  Oddly, it is not necessary to hide the
xml:lang attribute (at least not using xmllint/libxml2).

Note this commit includes a test case that (correctly) does not
validate.