fix element namespaces
The schema for XSD places elements in namespace
http://www.w3.org/2001/XMLSchema by default, but keeps attributes in no
namespace.  The placement of a default ns in the grammar element it
difficult to clearly distinguish these.

Remove the default namespace, explicitly qualify all element names, and
leave all attribute names unqualified so the ns attribute defaults to
the empty string per section 4.9 of the RELAX NG specification.