xsd10: refactor schema element to reduce duplication
Using interleave we can push the check for the targetNamespace down to
be in a group that only affects the import element.  This will simplify
XSD 1.1 support where there is a new composition element (xs:override)
that can precede schemaTop elements.

Note: This change changes the diagnostics from libxml2, which apparently
does not deal with interleave very well.  In this case the new
diagnostics are better.