cleanup: remove obsolete documentation
My understanding is the original intent of this schema was to attempt to
express the content model of XML Schema Definition within RELAX NG.
Most comments carried information about how XSD worked, and how the
RELAX NG implementation expressed them.

In its current use as purely validation for XSD with RELAX NG-capable
validators, it is more important that there be a clear correlation
between the RELAX NG schema and the normative Schema for Schemas.
Documentation should describe points of difference between the
recognized langauges, not redundantly summarize what XSD is intended to
recognize.